refactor: Move Grid3D and related components to the shared folder
