+++ src/aux-server/aux-web/aux-player/interaction/PlayerInteractionManager.ts
@@ -68,7 +68,7 @@ import {
 import { PlayerEmptyClickOperation } from './ClickOperation/PlayerEmptyClickOperation';
 import { PlayerGame } from '../scene/PlayerGame';
 import { DimensionGroup3D } from '../../shared/scene/DimensionGroup3D';
-import { Grid3D } from '../Grid3D';
+import { Grid3D } from '../../shared/scene/Grid3D';
 import { PlayerPageSimulation3D } from '../scene/PlayerPageSimulation3D';
 import { PlayerSimulation3D } from '../scene/PlayerSimulation3D';
 import { InventorySimulation3D } from '../scene/InventorySimulation3D';
