+++ src/aux-server/aux-web/shared/scene/BoundedGrid3D.spec.ts
