+++ src/aux-server/aux-web/shared/scene/CompoundGrid3D.spec.ts
