+++ src/aux-server/aux-web/shared/scene/Grid3D.ts
